Extract power policy mocks into `power-policy-interface-test-mocks` crate
Move PSU and charger mock code from power-policy-service/tests/common/mock.rs
into a standalone power-policy-interface-test-mocks crate with separate psu and
charger modules. Update workspace configuration, CI workflow exclusions, and
test imports accordingly.
