+++ power-policy-interface-test-mocks/src/lib.rs
@@ -1,0 +1,5 @@
+#![allow(clippy::unwrap_used)]
+#![allow(clippy::expect_used)]
+
+pub mod charger;
+pub mod psu;
