+++ power-policy-service/Cargo.toml
@@ -30,6 +30,7 @@ tokio = { workspace = true, features = ["rt", "macros", "time"] }
 env_logger = "0.11.8"
 log = { workspace = true }
 embedded-batteries-async = { workspace = true }
+power-policy-interface-test-mocks = { workspace = true }
 # TODO: figure out why enabling the log feature here causes running tests at the workspace level to fail to compile
 # Uncomment this line to enable log output in tests
 # power-policy-service = { workspace = true, features = ["log"] }
